+++ include/altacore/det/type.hpp
@@ -21,7 +21,10 @@ namespace AltaCore {
         static std::shared_ptr<Type> getUnderlyingType(AST::ExpressionNode* expression);
 
         bool isNative = true;
+        bool isFunction = false;
         NativeType nativeTypeName = NativeType::Integer;
+        std::shared_ptr<Type> returnType;
+        std::vector<std::shared_ptr<Type>> parameters;
 
         /**
          * follows the same format as `AltaCore::AST::Type::modifiers`
@@ -52,6 +55,7 @@ namespace AltaCore {
         bool isCompatiableWith(const Type& other);
 
         Type(NativeType nativeTypeName, std::vector<uint8_t> modifiers = {});
+        Type(std::shared_ptr<Type> returnType, std::vector<std::shared_ptr<Type>> parameters, std::vector<uint8_t> modifiers = {});
 
         // operator for `isCompatiableWith`
         bool operator %(const Type& other);